refactor. Upgraded to eleventy 3.0.0 and refreshed base layouts slightly.
