+++ package.json
@@ -2,7 +2,7 @@
   "name": "dw-website",
   "private": true,
   "description": "The personal and somewhat professional website for David Wesst, a.k.a. DW, a.k.a. Wessty.",
-  "version": "10.5.1",
+  "version": "11.0.0-alpha",
   "author": "David Wesst <david@cocobokostudios.com>",
   "homepage": "https://github.com/davidwesst/website#readme",
   "license": "MIT",
@@ -28,21 +28,22 @@
     "debug:eleventy": "DEBUG=Eleventy* eleventy",
     "generate-tags": "./tools/tag-helper.js 'src/blog' 'src/_data/content-tags.json'",
     "prestart": "npm run build",
-    "start": "swa start --output-location ./dist --swa-config-location ./",
+    "start": "http-server ./dist --port 8081",
     "test:workflow": "./bin/act -j build_and_deploy"
   },
   "devDependencies": {
     "@11ty/eleventy": "^3.0.0",
     "@11ty/eleventy-fetch": "^3.0.0",
-    "@11ty/eleventy-img": "^3.1.0",
-    "@11ty/eleventy-plugin-rss": "^1.2.0",
-    "@11ty/eleventy-plugin-webc": "^0.11.1",
+    "@11ty/eleventy-img": "^5.0.0",
+    "@11ty/eleventy-plugin-rss": "^2.0.2",
+    "@11ty/eleventy-plugin-webc": "^0.11.2",
     "concurrently": "^7.4.0",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.7.1",
     "date-fns": "^2.29.3",
     "dotenv-webpack": "^8.0.1",
     "gray-matter": "^4.0.3",
+    "http-server": "^14.1.1",
     "markdown-it": "^13.0.1",
     "markdown-it-eleventy-img": "^0.10.1",
     "markdown-it-emoji": "^2.0.2",
@@ -62,5 +63,6 @@
   "dependencies": {
     "@microsoft/applicationinsights-web": "^2.8.8",
     "sharp": "^0.31.1"
-  }
+  },
+  "type": "module"
 }
