+++ src/_data/site.json
@@ -1,5 +1,5 @@
 {
 	"name": "David Wesst",
 	"author": "David Wesst",
-	"url": "https://www.davidwesst.com"
+	"url": "https://davidwesst.com"
 }
