--- src/_includes/components/cookie-dialogue.html
@@ -1,11 +1,0 @@
-<dialog 
-    id="cookieDialogue"
-	class="dialogue"
-    >
-	<p>This site uses cookies to track usage in order to help improve the user experience. By clicking "Accept", you consent to our use of cookies.</p>
-	<form method="dialog">
-		<button id="btnCookieAccept" class="btn-square" onClick="acceptCookies(true)">Accept</button>
-		<button id="btnCookieReject" class="btn-square" onClick="acceptCookies(false)">Reject</button>
-		<a href="/about#privacy_statement">Privacy Statement</a>
-	</form>
-</dialog>
