--- src/_includes/components/event-item.webc
@@ -1,10 +1,0 @@
-<li>
-    <a :href="href">
-        <h3 class="event-item-title" @text="title"></h3>
-    </a>
-    <article>
-        <div @text="location"></div>
-        <time :datetime="startDate"
-            @text="formatDate(startDate)"></time>
-    </article>
-</li>
