--- src/_includes/components/event-list.webc
@@ -1,63 +1,0 @@
-<script webc:setup>
-    const sortFunc = (a, b) => new Date(b.data.duration.start) - new Date(a.data.duration.start);
-    const eventItems = Array.prototype.sort.call(collections.event, sortFunc);
-
-    const upcomingEvents = Array.prototype.filter.call(eventItems, (item) => { 
-        return Date.now() < new Date(item.data.duration.start);
-    });
-    //console.dir(upcomingEvents);
-    const upcomingEventSessionIds = [].concat(...upcomingEvents.map((item) => {
-        return [].concat(...item.data.talks.map(i => i.id));
-    }));
-    //console.dir(upcomingEventSessionIds);
-
-    const completedEvents = Array.prototype.filter.call(eventItems, (item) => { 
-        return Date.now() > new Date(item.data.duration.start); 
-    });
-    //console.dir(completedEvents.length);
-    const completedEventSessionIds = [].concat(...completedEvents.map((item) => {
-        return [].concat(...item.data.talks.map(i => i.id));
-    }));
-    //console.dir(completedEvents);
-</script>
-
-<section webc:if="talkFilter !== undefined">
-
-    <script webc:type="js" webc:root>
-        const filteredCompletedEvents = completedEvents.filter(obj => 
-            obj.data.talks.some(talk => talk.id === talkFilter)
-        );
-        const output = [];
-        filteredCompletedEvents.map(item => {
-            output.push(`<li><a href="${item.url}">${item.data.title}</a></li>`);
-        });
-        `<menu class="content-list">${output.join("")}</menu>`;
-    </script>
-</section>
-<section webc:else>
-    <article>
-        <h3>Upcoming Events</h3>
-        <menu class="content-list" webc:if="upcomingEvents.length > 0" >
-            <event-item webc:for="(value, index) of upcomingEvents" 
-            :href="value.url"
-            :@title="value.data.title"
-            :@start-date="value.data.duration.start"
-            :@location="`${value.data.location.city}, ${value.data.location.provinceState} (${value.data.location.country})`">
-        </event-item>
-        </menu>
-        <menu class="content-list" webc:else>
-            There are no upcoming events scheduled at this time.
-        </menu>
-    </article>
-    <article>
-        <h3>Completed Events</h3>
-        <menu class="content-list">
-            <event-item webc:for="(value, index) of completedEvents"
-                :href="value.url"
-                :@title="value.data.title"
-                :@start-date="value.data.duration.start"
-                :@location="`${value.data.location.city}, ${value.data.location.provinceState} (${value.data.location.country})`">
-            </event-item>
-        </menu>
-    </article>
-</section>
