+++ src/_includes/components/site-header.webc
@@ -7,11 +7,11 @@
             <a href="/">Home</a>
             <a href="/about">About</a>
             <a href="/blog">Blog</a>
-            <a href="/events">Events</a>
+            <!-- <a href="/events">Events</a> -->
             <a href="/talks">Talks</a>
             <!-- <a href="/gamelog">Gamelog</a> -->
-            <a href="/rememberthehuman">Remember The Human</a>
-            <a href="/cocobokostudios">Cocoboko Studios</a>
+            <!-- <a href="/remember-the-human">Remember The Human</a> -->
+            <a href="/cocoboko-studios">Cocoboko Studios</a>
         </nav>
     </div>
 </header>
