+++ src/_layouts/base.html
@@ -6,7 +6,7 @@
         <meta http-equiv="X-UA-Compatible" content="ie=edge" />
 		<meta name="generator" content="{{ eleventy.generator }}">
 		
-        {% assign page_url = "https://www.davidwesst.com" | append: page.url %}
+        {% assign page_url = "https://davidwesst.com" | append: page.url %}
         {% assign page_description = description %}
         {% assign page_type = type | default: "website" %}
 		{% assign page_image_height = image_height | default: 630 %}
@@ -30,28 +30,30 @@
         <meta name="tag" property="og:article:tag" content="{{ tag }}" />
         {% endfor %}
         {% endif %}
+
+        {% comment %}
         {% if image %}
 		{% capture page_image %}
 		    {% image_uri image %}
 		{% endcapture %}
-		<meta name="image" property="og:image" content="{{ page_image | strip |  prepend: 'https://www.davidwesst.com' }}" />
-		<meta property="og:image:secure_url" content="{{ page_image | strip | prepend: 'https://www.davidwesst.com' }}" />
+		<meta name="image" property="og:image" content="{{ page_image | strip |  prepend: 'https://davidwesst.com' }}" />
+		<meta property="og:image:secure_url" content="{{ page_image | strip | prepend: 'https://davidwesst.com' }}" />
 		<meta property="og:image:alt" content="{{ image_alt }}" />
 		{% else %}
-		<meta name="image" property="og:image" content="https://www.davidwesst.com/static/default_og.png" />
+		<meta name="image" property="og:image" content="https://davidwesst.com/static/default_og.png" />
         <meta property="og:image:alt" content="A pixel art version of David Wesst with glasses in a circle, looking to the top right towards text that is his name and the subtitle 'Code. Creativity. Cool Stuff'." />
 		{% endif %}
         <meta property="og:image:type" content="{{ page_image_mimetype }}" />
         <meta property="og:image:height" content="{{ page_image_height }}" />
         <meta property="og:image:width" content="{{ page_image_width }}" />
+        {% endcomment %}
 
 		<script defer src="/assets/main.js"></script>
     </head>
     <body>
-        {% render "components/site-header.html" %}
+        {% render "components/site-header.webc" %}
         <main class="bg-logo">
             {{ content }}
         </main>
-		{% render "components/cookie-dialogue.html" %}
     </body>
 </html>
