--- src/_layouts/event.webc
@@ -1,31 +1,0 @@
----
-layout: "base.html"
----
-<header>
-    <h1 @text="title"></h1>
-    <div class="subTitle">
-        <time :datetime="duration.start" @text="formatDate(duration.start)"></time> - <time :datetime="duration.end" @text="formatDate(duration.end)"></time>
-    </div>
-    <div class="subTitle" @text="`${location.city}, ${location.provinceState} (${location.country})`"></div>
-</header>
-<section webc:if="content" @html="content">
-    This element should not appear if there is no context in the event markdown file.
-</section> 
-<section>
-    <h2>Talks</h2>
-    <menu class="content-list">
-        <li webc:for="(t, index) of talks">
-            <a :href="`/talks/${t.id}`">
-                <h3 @text="`${t.title}`"></h3>
-            </a>
-            <h4>Files / Links</h4>
-            <menu webc:if="t.links">
-                <li webc:for="l of t.links">
-                    <a :href="l.url" @text="l.shortDescription"></a>
-                </li>
-            </menu>
-            <article webc:else>No files or links shared.</article>
-        </li>
-    </menu>
-</section>
-
