+++ src/about.html
@@ -1,0 +1,36 @@
+---
+layout: base.html
+permalink: /about/
+title: ABOUT // David Wesst
+description: Who is DW anyway?
+---
+<h1>
+    DW is David Wesst. 
+</h1>
+<section>
+    <p>I am a lot of things, but if you have to give me a label I would like to think I'm someone who likes to <em>code cool stuff</em>.</p>
+</section>
+<section id="privacy_statement">
+	<h2>Privacy Statement</h2>
+	<p>
+		I am a firm believer that your data is <em>your data</em>, not mine. 
+	</p>
+	<p>
+		That being said, you generate data just by using this website that can help me improve the site. This is why I use <a href="https://learn.microsoft.com/en-us/azure/azure-monitor/app/app-insights-overview">Application Insights</a> from Microsoft to monitor behaviour and performance of the site itself, along with how user interact with the site (assuming the user has opted-in for cookie usage).
+	</p>
+	<p>
+		The usage data is collected through the use of cookies, which are NOT enabled by default (although I would appreciate it if you opted-in).
+	</p>
+	<p>
+		For more information about how data is collected and managed with Application Insights, please read through the <a href="https://learn.microsoft.com/en-us/azure/azure-monitor/app/data-retention-privacy">Microsoft documentation about Application Insights</a>.
+</section>
+<section id="contact">
+    <h2>Contact a.k.a. Where to Find Me?</h2>
+    <nav class="social-links">
+        <a href="https://twitter.com/davidwesst">Twitter</a>
+        <a href="https://github.com/davidwesst">GitHub</a>
+        <a href="https://youtube.com/davidwesst">YouTube</a>
+        <a href="https://twitch.tv/davidwesst">Twitch</a>
+        <a href="https://ca.linkedin.com/in/davidwesst">LinkedIn</a>
+    </nav>
+</section>
