+++ src/cocoboko-studios.html
@@ -1,6 +1,6 @@
 ---
 layout: base.html
-permalink: /cocobokostudios/index.html
+permalink: /cocoboko-studios/index.html
 title: COCOBOKO STUDIOS // David Wesst
 description: Cocoboko Studios is a dream.
 ---
