--- src/events/events.11tydata.js
@@ -1,4 +1,0 @@
-module.exports = {
-    "layout": "event.webc",
-    "tags": [ "event" ]
-}
