--- src/events/informatixdiscover2023-winnipeg.md
@@ -1,18 +1,0 @@
----
-title: "Informatix Discover"
-url: "https://www.informatixdiscover.com"
-eventType: "conference"
-location:
-    city: "Winnipeg"
-    provinceState: "Manitoba"
-    country: "Canada"
-duration:
-    start: "2023-10-19"
-    end: "2023-10-19"
-talks:
-    - id: is-ai-ready-to-take-over-the-world
-      title: "Is AI Ready to Take Over the World?"
-      links:
-        - shortDescription: "Slides (PDF)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/is-ai-ready-to-take-over/is-ai-ready-to-take-over-the-world_informanixdiscover-2023.pdf"
----
