--- src/events/prairiedevcon2022-calgary.md
@@ -1,29 +1,0 @@
----
-title: "Prairie Dev Con 2022 - Calgary"
-url: "https://www.prairiedevcon.com"
-eventType: "conference"
-location:
-    city: "Calgary"
-    provinceState: "Alberta"
-    country: "Canada"
-duration:
-    start: "2022-11-28"
-    end: "2022-11-29"
-talks:
-    - id: "concensus-in-the-chaos"
-      title: "Consensus in the Chaos​: The Role of IT Architecture​"
-      links:
-        - shortDescription: "Slides (PPTX)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/concensus-in-the-chaos/Consensus%20in%20the%20Chaos%20(PrDC%202022)%20-%20Calgary.pptx"
-        - shortDescription: "Slides (PDF)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/concensus-in-the-chaos/Consensus%20in%20the%20Chaos%20(PrDC%202022)%20-%20Calgary.pdf"
-    - id: "cots-to-cloud"
-      title: "From Custom COTS to Cloud: A Case Study in Solution Architecture"
-      links:
-        - shortDescription: "Slides (PPTX)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/cots-to-cloud/From%20COTS%20to%20Cloud%20(PrDC%202022)%20-%20Calgary.pptx"
-        - shortDescription: "Slides (PDF)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/cots-to-cloud/From%20COTS%20to%20Cloud%20(PrDC%202022)%20-%20Calgary.pdf"
-
----
-
