--- src/events/prairiedevcon2022-winnipeg.md
@@ -1,28 +1,0 @@
----
-title: "Prairie Dev Con 2022 - Winnipeg"
-url: "https://www.prairiedevcon.com"
-eventType: "conference"
-location:
-    city: "Winnipeg"
-    provinceState: "Manitoba"
-    country: "Canada"
-duration:
-    start: "2022-11-07"
-    end: "2022-11-08"
-talks:
-    - id: "concensus-in-the-chaos"
-      title: "Consensus in the Chaos​: The Role of IT Architecture​"
-      links:
-        - shortDescription: "Slides (PPTX)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/concensus-in-the-chaos/Consensus in the Chaos (PrDC 2022) - Winnipeg.pptx"
-        - shortDescription: "Slides (PDF)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/concensus-in-the-chaos/Consensus in the Chaos (PrDC 2022) - Winnipeg.pdf"
-    - id: "cots-to-cloud"
-      title: "From Custom COTS to Cloud: A Case Study in Solution Architecture"
-      links:
-        - shortDescription: "Slides (PDF)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/cots-to-cloud/From COTS to Cloud (PrDC 2022) - Winnipeg.pptx"
-        - shortDescription: "Slides (PDF)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/cots-to-cloud/From COTS to Cloud (PrDC 2022) - Winnipeg.pdf"
-
----
