--- src/events/prairiedevcon2023-regina.md
@@ -1,24 +1,0 @@
----
-title: "Prairie Dev Con 2023 - Regina"
-url: "https://www.prairiedevcon.com"
-eventType: "conference"
-location:
-    city: "Regina"
-    provinceState: "Saskatchewan"
-    country: "Canada"
-duration:
-    start: "2023-10-16"
-    end: "2022-10-17"
-talks:
-    - id: is-ai-ready-to-take-over-the-world
-      title: "Is AI Ready to Take Over the World? Let's Discuss!"
-      links:
-        - shortDescription: "Slides (PDF)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/is-ai-ready-to-take-over/is-ai-ready-to-take-over-the-world_prdc-regina-2023.pdf"
-    - id: hack-the-it-governance-matrix
-      title: "Hack the IT Governance Matrix: How Solution Architects Can Drive Change Without the Red Pill"
-      links:
-        - shortDescription: "Slides (PDF)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/hack-the-it-governance-matrix/hack-the-it-governance-matrix_prdc-regina-2023.pdf"
----
-
