--- src/events/prairiedevcon2023-winnipeg.md
@@ -1,23 +1,0 @@
----
-title: "Prairie Dev Con 2023 - Winnipeg"
-url: "https://www.prairiedevcon.com"
-eventType: "conference"
-location:
-    city: "Winnipeg"
-    provinceState: "Manitoba"
-    country: "Canada"
-duration:
-    start: "2023-06-12"
-    end: "2023-06-13"
-talks:
-    - id: is-ai-ready-to-take-over-the-world
-      title: "Is AI Ready to Take Over the World?"
-      links:
-        - shortDescription: "Slides (PDF)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/is-ai-ready-to-take-over/is-ai-ready-to-take-over-the-world_prdc-winnipeg-2023.pdf"
-    - id: hack-the-it-governance-matrix
-      title: "Hack the IT Governance Matrix: How Solution Architects Can Drive Change Without the Red Pill"
-      links:
-        - shortDescription: "Slides (PDF)"
-          url: "https://davidwesstweb.blob.core.windows.net/talks/hack-the-it-governance-matrix/hack-the-it-governance-matrix_prdc-winnipeg-2023.pdf"
----
