--- src/events/scared2023-winnipeg.md
@@ -1,15 +1,0 @@
----
-title: "Scared S#!tless Starting a Business (2023)"
-url: "https://eventcamp.ca/event/it-s-scary-talking-about-technology-a-i/"
-eventType: "user-group"
-location:
-    city: "Winnipeg"
-    provinceState: "Manitoba"
-    country: "Canada"
-duration:
-    start: "2023-10-23"
-    end: "2023-10-27"
-talks:
-    - id: its-scary-using-new-ai-technology
-      title: "It's Scary Using New AI Technology"
----
