--- src/feeds/blog-feed.njk
@@ -1,37 +1,0 @@
----json
-{
-  "permalink": "/blog/feed.xml",
-  "eleventyExcludeFromCollections": true,
-  "metadata": {
-    "title": "Blog by DW",
-    "subtitle": "My ramblings about technology, life, and whatever else I come across in my adventure with code.",
-    "language": "en",
-    "url": "https://www.davidwesst.com/",
-    "author": {
-      "name": "David Wesst",
-      "email": "dw@davidwesst.com"
-    }
-  }
-}
----
-<?xml version="1.0" encoding="utf-8"?>
-<rss version="2.0" xmlns:dc="http://purl.org/dc/elements/1.1/" xml:base="{{ metadata.url }}" xmlns:atom="http://www.w3.org/2005/Atom">
-  <channel>
-    <title>{{ metadata.title }}</title>
-    <link>{{ metadata.url }}</link>
-    <atom:link href="{{ permalink | absoluteUrl(metadata.url) }}" rel="self" type="application/rss+xml" />
-    <description>{{ metadata.subtitle }}</description>
-    <language>{{ metadata.language }}</language>
-    {%- for post in collections.blog | reverse %}
-    {%- set absolutePostUrl = post.url | absoluteUrl(metadata.url) %}
-    <item>
-      <title>{{ post.data.title }}</title>
-      <link>{{ absolutePostUrl }}</link>
-      <description>{{ post.templateContent | htmlToAbsoluteUrls(absolutePostUrl) }}</description>
-      <pubDate>{{ post.date | dateToRfc822 }}</pubDate>
-      <dc:creator>{{ metadata.author.name }}</dc:creator>
-      <guid>{{ absolutePostUrl }}</guid>
-    </item>
-    {%- endfor %}
-  </channel>
-</rss>
