--- src/meta/robots.njk
@@ -1,8 +1,0 @@
----
-permalink: /robots.txt
-eleventyExcludeFromCollections: true
----
-Sitemap: https://www.davidwesst.com/sitemap.xml
-
-User-agent: *
-Disallow:
