--- src/meta/sitemap.njk
@@ -1,16 +1,0 @@
----
-permalink: /sitemap.xml
-eleventyExcludeFromCollections: true
----
-<?xml version="1.0" encoding="utf-8"?>
-<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-    {% for page in collections.all %}
-       {% if not page.data.draft %}
-        <url>
-            <loc>{{ site.url }}{{ page.url | url }}</loc>
-            <lastmod>{{ page.date.toISOString() }}</lastmod>
-            <changefreq>{{ page.data.changeFreq if page.data.changeFreq else "monthly" }}</changefreq>
-        </url>
-        {% endif %}
-    {% endfor %}
-</urlset>
