--- src/pages/gamelog-list.html
@@ -1,13 +1,0 @@
----
-layout: base.html
-permalink: /gamelog/
-title: GAMELOG // David Wesst
-description: Returning at some point in the future...
-changefreq: weekly
----
-<h1>
-    Gamelog
-</h1>
-<section>
-    This section will return soon. I still play video games, but the backend that generated this list needs some work. 
-</section>
